% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTidyGene.R
\name{scCEGs}
\alias{scCEGs}
\title{Generate a tidy list of conserved expressed genes}
\usage{
scCEGs(seu_obj, ident.1, grouping.var)
}
\arguments{
\item{seu_obj}{A Seurat object containing single-cell RNA-seq data.}

\item{ident.1}{The identity of the cluster to analyze.}

\item{grouping.var}{The variable to group by when finding conserved markers.}
}
\value{
A data frame of conserved markers across clusters.
}
\description{
Generate a tidy list of conserved expressed genes
}
\examples{
# Example usage will be added in a future release.
}
