% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTidyGene.R
\name{scDEGs}
\alias{scDEGs}
\title{Generate a tidy list of differentially expressed genes}
\usage{
scDEGs(
  seu_obj,
  ident.1,
  ident.2,
  group.by,
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25
)
}
\arguments{
\item{seu_obj}{A Seurat object containing single-cell RNA-seq data.}

\item{ident.1}{The identity of the first condition to compare.}

\item{ident.2}{The identity of the second condition to compare.}

\item{group.by}{The variable to group by when performing differential expression analysis.}

\item{only.pos}{Logical, if TRUE, only positive markers are returned.}

\item{min.pct}{Minimum percentage of cells expressing the gene.}

\item{logfc.threshold}{Minimum log-fold change threshold.}
}
\value{
A data frame of differentially expressed genes across clusters.
}
\description{
Generate a tidy list of differentially expressed genes
}
\examples{
# Example usage will be added in a future release.
# result <- scDEGs(seu_obj, "ConditionA", "ConditionB", group.by = "group_var")
}
